Implemented referer check in configure service block to deny direct access, refactored check form nonce to pass in nonce key and value dynamically and renamed classes + methods

+++ Classes/Blocks/external-services.configure-service-block.php
@@ -4,17 +4,26 @@
 namespace ExternalServices\Classes\Blocks;
 
 use ExternalServices\Classes\Models\ES_Temp_Model;
+use http\Exception\RuntimeException;
 
 class Configure_Service_Block extends Block_Base {
 
 	public function __construct( array $data = null ) {
-		$tempModel = new ES_Temp_Model();
+		if (isset($_SERVER['HTTP_REFERER']) && $_SERVER['HTTP_REFERER'] == admin_url('admin.php?page=external-services-add')) {
+			$tempModel = new ES_Temp_Model();
 
-		$tempData = $tempModel->get();
-		$data = unserialize(base64_decode($tempData[0]->data));
-		$tempModel->delete(null);
+			$tempData = $tempModel->get();
+			$data = unserialize(base64_decode($tempData[0]->data));
+			$tempModel->delete(null);
 
-		parent::__construct($data);
+			parent::__construct($data);
+		} else {
+			wp_die('Direct access to this page is prohibited.');
+		}
+	}
+
+	public function getPageKey(): String {
+		return 'configure_service';
 	}
 
 	public function getFirstIndexOfConnectionData(): ?Object {
